# m_fleet/python-flask-server-generated/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: VTS API
  description: This API Provides dynamic communication with a down stream AMR Subsystem
    (MAS).
  version: winpg-1.2.4
servers:
- url: /
...
paths:
...
  /loads:
    get:
      tags:
      - Load Units
      summary: Gets load information from AMR subsystem for the specified type and
        load host ids
      description: "This API is typically used to retrieve current location of the\
        \ load. \nIf load does not exist for a given load host id, no record is returned\
        \ for that load host id."
      operationId: loads_get
      parameters:
      - name: loadHostIds
        in: query
        description: A comma separated list of load host Ids to filter the loads;
          At least one load host id must be provided
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: type
        in: query
        description: The type of load to be retrieved
        required: true
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/LoadTypes'
      responses:
        "200":
          description: The list of load
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Load'
                x-content-type: application/json
        "500":
          description: Retrieval of loads from AMR subsystem failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.load_units_controller
    post:
      tags:
      - Load Units
      summary: Adds the load unit into the warehouse
      description: Adds the load unit into the warehouse to carry out the fulfillment
        order.
      operationId: loads_post
      requestBody:
        description: The new load unit to add to the warehouse
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Load'
        required: true
      responses:
        "200":
          description: The new load unit added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
        "400":
          description: If the data in the load unit is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        "409":
          description: The specified load unit conflicts with a load unit that already
            exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        "500":
          description: Addition of the load unit failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.load_units_controller
    delete:
      tags:
      - Load Units
      summary: Removes the load unit from the warehouse
      description: Removes the load unit from the warehouse
      operationId: loads_delete
      requestBody:
        description: The load unit to be removed from to the warehouse
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Load'
        required: true
      responses:
        "200":
          description: The removed load unit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Load'
        "400":
          description: If the data in the load unit is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        "409":
          description: The specified load unit doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
        "500":
          description: Removal of the load unit failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpErrorResponse'
      x-openapi-router-controller: swagger_server.controllers.load_units_controller
...
components:
  schemas:
...
    Dimensions:
      type: object
      properties:
        length:
          type: number
          description: Defines the length of the object in the length units of the
            specified units system
          format: double
        width:
          type: number
          description: Defines the width of the object in the length units of the
            specified units system
          format: double
        height:
          type: number
          description: Defines the height of the object in the length units of the
            specified units system
          format: double
        weightLimit:
          type: number
          description: Defines the weight limit of the load in the weight units of
            the specified units system
          format: double
        weight:
          type: number
          description: Defines the weight of the load in the weight units of the specified
            units system
          format: double
        unitSystem:
          type: string
          description: The unit system used as the units for all the measurements
            in the object
          allOf:
          - $ref: '#/components/schemas/UnitSystems'
      description: Defines a structure to store the dimensions of an object
...
    HttpErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: The statusCode of HTTP errors reported back to the client
          format: int32
          readOnly: true
          example: 500
        detail:
          type: string
          description: The detail of HTTP errors reported back to the client
          readOnly: true
          example: Internal Server Error
        instance:
          type: string
          description: The URI response reported back to the client
          readOnly: true
          example: /fulfillment-missions
      description: The response body for HTTP errors reported back to the client
    Load:
      type: object
      properties:
        loadHostId:
          type: string
          description: Client identifier for this load. Mandatory for Pick and Drop
            task types.
          example: "1001"
        loadTypeId:
          type: string
          description: The ID of the type of load
        locationId:
          type: string
          description: The ID of the location of the load
          example: AGV.DP02.06.0.RL0610050.4036.4
        count:
          type: integer
          description: The count of object within the load
          format: int32
          example: 0
        type:
          type: string
          description: The type of load
          example: Pallet
          allOf:
          - $ref: '#/components/schemas/LoadTypes'
        dimensions:
          type: object
          description: The dimensions of the load
          allOf:
          - $ref: '#/components/schemas/Dimensions'
        customMetaData:
          type: array
          description: A collection of custom meta data to store with the load
          items:
            $ref: '#/components/schemas/MetaData'
        loadId:
          type: string
          description: The VTS ID of the load
          readOnly: true
          example: 9c3b1830-513c-33b8-bc4b-76635d32e692
        status:
          type: string
          description: The status of the load
          readOnly: true
          example: Enroute
          allOf:
          - $ref: '#/components/schemas/LoadStates'
      description: Defines the load that can be moved within the warehouse
      example:
        loadHostId: "1001"
        loadId: 9c3b1830-513c-33b8-bc4b-76635d32e692
        loadTypeId: loadTypeId
        locationId: AGV.DP02.06.0.RL0610050.4036.4
        count: 0
        customMetaData:
        - value: value
          key: key
        - value: value
          key: key
        type: Pallet
        dimensions: ""
        status: Enroute
...
    LoadStates:
      type: string
      description: Defines the states of a load in the warehouse
      enum:
      - Stored
      - Enroute
      - Discarded
    LoadTypes:
      type: string
      description: "<pre>The types of loads supported by the VTS system<br><br>Pallet\
        \ \t- The Pallet<br><br>Tote\t\t- The Tote<br><br>Case\t\t- The Case<br><br>Rack\t\
        \t- The Rack<br><br></pre>"
      enum:
      - Pallet
      - Tote
      - Case
      - Rack
...
    MetaData:
      type: object
      properties:
        key:
          type: string
          description: The key to identify the meta data
        value:
          type: string
          description: The data to store
      description: Defines a structure to store custom data
      example:
        value: value
        key: key
...
    UnitSystems:
      type: string
      description: The units used for measurements in VTS requests
      enum:
      - Metric
      - British
      - USA
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_SecurityScheme